Moved, renamed configuration and fixed issue when user cancel dialog file
