+++ addie/menu/file/configuration/import_configuration.py
@@ -2,7 +2,7 @@ from __future__ import (absolute_import, division, print_function)
 import configparser
 import os
 
-from addie.configuration.config_file_name_handler import ConfigFileNameHandler
+from addie.menu.file.configuration.config_file_name_handler import ConfigFileNameHandler
 from addie.utilities.conversion import str2bool
 from addie.utilities.gui_handler import GuiHandler
 from addie.autoNOM.step1_gui_handler import Step1GuiHandler
